+++ pAudio/code/share/www/public/js/main.js
@@ -3,15 +3,18 @@
     This file is part of 'pAudio', a PC based personal audio system.
 */
 
+import config from '../config.json' with { type: 'json' };
 import * as mc  from "./miscel.js";
 import * as hl  from "./highlight.js";
 import * as hlp from "./help.js";
 
+document.title = 'pAudio ' + config.loudspeaker;
 
 //////// GLOBAL VARIABLES ////////
 
 var POLLING_DELAY       = 1000;     // millisec
-var STATE               = {};
+
+var STATE               = { "loudspeaker": config.loudspeaker };
 
 var player_info         = {};
 
@@ -186,7 +189,7 @@ function fill_in_page_statics(){
     }
 
 
-    manage_main_cside( ':: pAudio :: ' + STATE.loudspeaker );
+    manage_main_cside( ':: pAudio : ' + STATE.loudspeaker );
 
 
     // updates level cell info with ref_SPL
@@ -210,7 +213,7 @@ async function init(){
     function preparare_UI(){
         document.getElementById("but_help").style.display = 'none';
         document.getElementById("but_restart").style.display = 'block';
-        document.getElementById("main_cside").innerHTML = ":: pAudio not connected ::";
+        document.getElementById("main_cside").innerHTML = ":: pAudio : " + STATE.loudspeaker + " ::<br>(not connected)";
         document.getElementById("levelInfo").innerHTML = "--";
     }
 
@@ -827,7 +830,7 @@ function manage_main_cside( msg = '' ){
         msg = '( sleeping )';
 
     } else if( ! CamillaDSP_is_ready() ) {
-        msg = 'DSP unloaded, needs restart';
+        msg = ':: pAudio : ' + STATE.loudspeaker + ' ::\n(DSP unloaded, needs restart)';
         document.getElementById("but_restart").style.display = "inline-block";
         document.getElementById("but_help").style.display = "none";
         mc.player_info_clear();