+++ doc/NEWS.md
@@ -28,3 +28,4 @@
 - Fix timeout for the pAudio server to be alive on slow machines.
 - Fix libresport reading events for metadata retrieving.
 - Add shairplay-sync plugin for Airplay receiving.
+- All custom configurations now under the pAudio/config/ folder